Move getLabel to helpers
