+++ lib/commands/validate.js
@@ -1,6 +1,6 @@
 const {readFile} = require('fs-extra')
 const chalk = require('chalk')
-const {validate, getLabel} = require('..')
+const {validate, getLabel} = require('../helpers')
 
 module.exports = {
   command: 'validate [options] <file>',
