+++ lib/helpers.js
@@ -1,4 +1,5 @@
 const profiles = require('./schema/profiles')
+const errorLabels = require('./schema/error-labels')
 
 function getErrorLevel(profileName, code) {
   const profile = profiles[profileName]
@@ -14,4 +15,29 @@ function getErrorLevel(profileName, code) {
   return 'I'
 }
 
-module.exports = {getErrorLevel}
+const endsWithErrorLabels = {}
+
+for (const c of Object.keys(errorLabels)) {
+  if (c.includes('*') && c.startsWith('*.')) {
+    endsWithErrorLabels[c.slice(2)] = submittedCode => {
+      const [value] = submittedCode.split('.')
+      return errorLabels[c].replace('{}', value)
+    }
+  }
+}
+
+function getLabel(code) {
+  if (code in errorLabels) {
+    return errorLabels[code]
+  }
+
+  const endsWithCandidate = Object.keys(endsWithErrorLabels).find(pattern => code.endsWith(pattern))
+
+  if (endsWithCandidate) {
+    return endsWithErrorLabels[endsWithCandidate](code)
+  }
+
+  return code
+}
+
+module.exports = {getErrorLevel, getLabel}
